+++ src/ext_widgets/image_animation/image_animation.h
@@ -39,7 +39,6 @@ BEGIN_C_DECLS
  * 在xml中使用"image\_animation"标签创建图片动画控件。如：
  *
  * ```xml
- * <image_animation image="ani" sequence="123456789abc" auto_play="true" interval="50"/>
  * <image_animation image="ani" start_index="1" end_index="9" auto_play="true" interval="50"
  * delay="100"/>
  * ```
@@ -53,7 +52,7 @@ BEGIN_C_DECLS
  * image_animation = image_animation_create(win, 10, 10, 200, 200);
  * image_animation_set_image(image_animation, "ani");
  * image_animation_set_interval(image_animation, 50);
- * image_animation_set_sequence(image_animation, "123456789abc");
+ * image_animation_set_range_sequence(image_animation, 1, 9);
  * image_animation_play(image_animation);
  * ```
  *
@@ -77,6 +76,7 @@ typedef struct _image_animation_t {
    * @property {char*} sequence
    * @annotation ["set_prop","get_prop","readable","persitent","design","scriptable"]
    * 播放的序列，字符可选值为:0-9,a-z,A-Z。
+   *
    */
   char* sequence;
   /**
@@ -103,6 +103,18 @@ typedef struct _image_animation_t {
    * 是否自动播放。
    */
   bool_t auto_play;
+  /**
+   * @property {bool_t} unload_after_paint
+   * @annotation ["set_prop","get_prop","readable","persitent","design","scriptable"]
+   * 绘制完成后unload图片，以释放内存空间。
+   */
+  bool_t unload_after_paint;
+  /**
+   * @property {char*} format
+   * @annotation ["set_prop","get_prop","readable","persitent","design","scriptable"]
+   * 索引到图片名转换时的格式，缺省为"%s%d"。
+   */
+  char* format;
   /**
    * @property {uint32_t} interval
    * @annotation ["set_prop","get_prop","readable","persitent","design","scriptable"]
@@ -193,7 +205,8 @@ ret_t image_animation_set_auto_play(widget_t* widget, bool_t auto_play);
 
 /**
  * @method image_animation_set_sequence
- * 设置播放序列。比如image为"fire"，sequence为"123", 将依次播放"fire1", "fire2", "fire3"。
+ * 设置播放序列。比如image为"fire"，sequence为"12223", 将依次播放"fire1", "fire2", "fire2", "fire2",
+ * "fire3"。
  * @annotation ["scriptable"]
  * @param {widget_t*} widget image_animation对象。
  * @param {const char*} sequence 播放序列。
@@ -204,7 +217,8 @@ ret_t image_animation_set_sequence(widget_t* widget, const char* sequence);
 
 /**
  * @method image_animation_set_range_sequence
- * 设置播放序列。比如image为"fire"，max_nr为100, 将依次播放"fire0", ..., "fire99"。
+ * 设置播放序列。比如image为"fire"，start_index为0, end_index为99, 将依次播放"fire0", ...,
+ * "fire99"。
  * @annotation ["scriptable"]
  * @param {widget_t*} widget image_animation对象。
  * @param {uint32_t} start_index 图片起始序数。
@@ -246,6 +260,36 @@ ret_t image_animation_stop(widget_t* widget);
 ret_t image_animation_pause(widget_t* widget);
 
 /**
+ * @method image_animation_set_format
+ * 设置生成图片名的格式。
+ *
+ * XXX:生成图片名时，第一个参数是图片名前缀，第二个是序数，只能在此前提下设置格式。
+ *
+ * ```
+ *  const char* format = image_animation->format ? image_animation->format : "%s%d";
+ *  tk_snprintf(name, TK_NAME_LEN, format, image_animation->image, image_animation->index);
+ * ```
+ *
+ * @annotation ["scriptable"]
+ * @param {widget_t*} widget image_animation对象。
+ * @param {const char*} format 格式。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t image_animation_set_format(widget_t* widget, const char* format);
+
+/**
+ * @method image_animation_set_unload_after_paint
+ * 设置绘制完成后unload图片，以释放内存空间。
+ * @annotation ["scriptable"]
+ * @param {widget_t*} widget image_animation对象。
+ * @param {bool_t} unload_after_paint 是否绘制完成后unload图片。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t image_animation_set_unload_after_paint(widget_t* widget, bool_t unload_after_paint);
+
+/**
  * @method image_animation_cast
  * 转换为image_animation对象(供脚本语言使用)。
  * @annotation ["cast", "scriptable"]
@@ -261,12 +305,15 @@ widget_t* image_animation_cast(widget_t* widget);
 #define IMAGE_ANIMATION_PROP_END_INDEX "end_index"
 #define IMAGE_ANIMATION_PROP_INTERVAL "interval"
 #define IMAGE_ANIMATION_PROP_AUTO_PLAY "auto_play"
+#define IMAGE_ANIMATION_PROP_UNLOAD_AFTER_PAINT "unload_after_paint"
 
 #define WIDGET_TYPE_IMAGE_ANIMATION "image_animation"
 #define IMAGE_ANIMATION(widget) ((image_animation_t*)(widget))
 
 /*public for test*/
 ret_t image_animation_update(widget_t* widget);
+ret_t image_animation_get_image_name(image_animation_t* image_animation,
+                                     char name[TK_NAME_LEN + 1]);
 
 END_C_DECLS