+++ src/tkc/value.c
@@ -365,6 +365,7 @@ ret_t value_copy(value_t* dst, const value_t* src) {
   }
 
   memcpy(dst, src, sizeof(value_t));
+  dst->free_handle = FALSE;
 
   return RET_OK;
 }
@@ -376,18 +377,18 @@ ret_t value_deep_copy(value_t* dst, const value_t* src) {
 
   switch (dst->type) {
     case VALUE_TYPE_STRING: {
-      dst->free_handle = TRUE;
       dst->value.str = tk_strdup(src->value.str);
+      dst->free_handle = dst->value.str != NULL;
       break;
     }
     case VALUE_TYPE_WSTRING: {
-      dst->free_handle = TRUE;
       dst->value.wstr = tk_wstrdup(src->value.wstr);
+      dst->free_handle = dst->value.wstr != NULL;
       break;
     }
     case VALUE_TYPE_OBJECT: {
-      dst->free_handle = TRUE;
       object_ref(dst->value.object);
+      dst->free_handle = dst->value.object != NULL;
       break;
     }
     default: