+++ tests/image_animation_test.cc
@@ -45,6 +45,11 @@ TEST(ImageAnimation, basic) {
   ASSERT_EQ(widget_get_prop(w, IMAGE_ANIMATION_PROP_AUTO_PLAY, &v2), RET_OK);
   ASSERT_EQ(value_bool(&v1), value_bool(&v2));
 
+  value_set_bool(&v1, TRUE);
+  ASSERT_EQ(widget_set_prop(w, IMAGE_ANIMATION_PROP_UNLOAD_AFTER_PAINT, &v1), RET_OK);
+  ASSERT_EQ(widget_get_prop(w, IMAGE_ANIMATION_PROP_UNLOAD_AFTER_PAINT, &v2), RET_OK);
+  ASSERT_EQ(value_bool(&v1), value_bool(&v2));
+
   widget_destroy(w);
 }
 
@@ -84,6 +89,38 @@ TEST(ImageAnimation, range) {
   ASSERT_EQ(ia->index, 101);
   image_animation_update(w);
   ASSERT_EQ(ia->index, 102);
+
+  widget_destroy(w);
+}
+
+TEST(ImageAnimation, format) {
+  char name[TK_NAME_LEN + 1];
+  widget_t* w = image_animation_create(NULL, 10, 20, 30, 40);
+  image_animation_t* ia = IMAGE_ANIMATION(w);
+
+  ASSERT_EQ(widget_set_prop_str(w, WIDGET_PROP_FORMAT, "%s_%04d"), RET_OK);
+  ASSERT_STREQ(widget_get_prop_str(w, WIDGET_PROP_FORMAT, ""), "%s_%04d");
+
+  ASSERT_EQ(image_animation_set_format(w, "%s_%02d"), RET_OK);
+  ASSERT_STREQ(widget_get_prop_str(w, WIDGET_PROP_FORMAT, ""), "%s_%02d");
+
+  ia->index = 1;
+  ASSERT_EQ(image_animation_set_image(w, "ani"), RET_OK);
+  ASSERT_EQ(image_animation_get_image_name(ia, name), RET_OK);
+  ASSERT_STREQ(name, "ani_01");
+
+  ia->index = 11;
+  ASSERT_EQ(image_animation_get_image_name(ia, name), RET_OK);
+  ASSERT_STREQ(name, "ani_11");
+
+  ASSERT_EQ(image_animation_set_format(w, "%s_%x"), RET_OK);
+  ia->index = 11;
+  ASSERT_EQ(image_animation_get_image_name(ia, name), RET_OK);
+  ASSERT_STREQ(name, "ani_b");
+
+  ia->index = 12;
+  ASSERT_EQ(image_animation_get_image_name(ia, name), RET_OK);
+  ASSERT_STREQ(name, "ani_c");
 
   widget_destroy(w);
 }