+++ docs/changes.md
@@ -1,4 +1,25 @@
 # 最新动态
+* 2019/02/25
+  * 完善image\_animation，修改#85
+  * 重构SConstruct，把公共代码放到awtk\_config.py，可供其它项目使用。
+
+* 2019/02/22
+  * 完善控件文档(感谢俊杰提供补丁)。
+  * 更新IDL。
+  * 修改awtk-js代码产生器中生成cast函数的问题。
+
+* 2019/02/21
+  * 增加object\_copy\_prop
+  * 修改value\_copy/value\_deep\_copy的BUG。
+  * 修改named\_value\_set\_value的BUG。
+  * 完善布局文档(感谢俊杰提供补丁)。
+  * 支持动态设置屏保时间。
+
+* 2019/02/20
+  * 按钮增加enable\_long\_press属性用于启用长按事件，触发长按事件后不再触发点击事件。
+  * 增加timer\_reset函数。
+  * 增加屏保支持。
+
 * 2019/02/19
   * add darray\_tail/darray\_head。